+++ modules/util/portal-tools-data-partitioning-sql-builder/src/main/java/com/liferay/portal/tools/data/partitioning/sql/builder/internal/command/DeleteExportProcessCommand.java
@@ -12,11 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.tools.data.partitioning.sql.builder.internal.command.impl;
+package com.liferay.portal.tools.data.partitioning.sql.builder.internal.command;
 
 import com.liferay.portal.tools.data.partitioning.sql.builder.exporter.DBExporter;
 import com.liferay.portal.tools.data.partitioning.sql.builder.exporter.context.ExportContext;
-import com.liferay.portal.tools.data.partitioning.sql.builder.internal.command.BaseExportProcessCommand;
 
 import java.io.OutputStream;
 
@@ -25,10 +24,9 @@ import java.util.List;
 /**
  * @author Manuel de la Peña
  */
-public class InsertPartitionedExportProcessCommand
-	extends BaseExportProcessCommand {
+public class DeleteExportProcessCommand extends BaseExportProcessCommand {
 
-	public InsertPartitionedExportProcessCommand(
+	public DeleteExportProcessCommand(
 		long companyId, DBExporter dbExporter, List<String> tableNames,
 		ExportContext exportContext) {
 
@@ -37,19 +35,18 @@ public class InsertPartitionedExportProcessCommand
 
 	@Override
 	protected String getOutputFileName() {
-		return exportContext.getSchemaName() + "-" + companyId +
-			"-partitioned" + dbExporter.getOutputFileExtension();
+		return exportContext.getSchemaName() + "-" + companyId + "-delete.sql";
 	}
 
 	@Override
 	protected String getOutputFileName(String tableName) {
 		return exportContext.getSchemaName() + "-" + companyId + "-table-" +
-			tableName + dbExporter.getOutputFileExtension();
+			tableName + "-delete.sql";
 	}
 
 	@Override
 	protected void write(String tableName, OutputStream outputStream) {
-		dbExporter.write(companyId, tableName, outputStream);
+		dbExporter.writeDelete(companyId, tableName, outputStream);
 	}
 
 }
